Remove trueVerifier from main logic
This verifier implementation is only used in tests, so it should be defined there.
