+++ pkg/download/verifier_test.go
@@ -58,37 +58,3 @@ func TestSha256Verifier(t *testing.T) {
 		})
 	}
 }
-
-func TestTrueVerifier(t *testing.T) {
-	tests := []struct {
-		name      string
-		write     []byte
-		wantError bool
-	}{
-		{
-			name:      "test okay hash",
-			write:     []byte("hello world"),
-			wantError: false,
-		},
-		{
-			name:      "test wrong hash",
-			write:     []byte("HELLO WORLD"),
-			wantError: false,
-		},
-		{
-			name:      "test empty hash",
-			write:     []byte{},
-			wantError: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			v := NewInsecureVerifier()
-			_, _ = io.Copy(v, bytes.NewReader(tt.write))
-			if err := v.Verify(); (err != nil) != tt.wantError {
-				t.Errorf("NewInsecureVerifier().Write(%x).Verify() = %v, wantReader %v", tt.write, err, tt.wantError)
-				return
-			}
-		})
-	}
-}
